Solution cleanup
