+++ Dungeon.Net.MonsterCrawl/Card.cs
@@ -1,0 +1,17 @@
+using System.Text.Json.Serialization;
+
+namespace Dungeon.Net.MonsterCrawl;
+
+public enum CardType
+{
+    Monster,
+    Weapon,
+    Potion,
+}
+
+public class Card
+{
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public CardType CardType { get; init; }
+    public required int Value { get; init; }
+}
