+++ Dungeon.Net.MonsterCrawl/Dungeon.Net.MonsterCrawl.csproj
@@ -1,9 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="assets/*.json" CopyToOutputDirectory="PreserveNewest" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Dungeon.Net.Engine\Dungeon.Net.Engine.csproj" />
+  </ItemGroup>
 
 </Project>
