--- Dungeon.Net.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.3" />
-    <PackageReference Include="Spectre.Console" Version="0.54.0" />
-    <PackageReference Include="Spectre.Console.ImageSharp" Version="0.54.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="cards.json" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
-
-</Project>
