+++ Dungeon.Net.sln
@@ -1,19 +1,60 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dungeon.Net", "Dungeon.Net.csproj", "{7B696AA6-98DE-CB02-A519-584851008534}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dungeon.Net.Launcher", "Dungeon.Net.Launcher\Dungeon.Net.Launcher.csproj", "{CDEF222D-7FE3-494E-AA82-A6E505583A4B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dungeon.Net.Engine", "Dungeon.Net.Engine\Dungeon.Net.Engine.csproj", "{F2455A4D-2DDD-42E0-B640-0D8065AC5637}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dungeon.Net.MonsterCrawl", "Dungeon.Net.MonsterCrawl\Dungeon.Net.MonsterCrawl.csproj", "{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B696AA6-98DE-CB02-A519-584851008534}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B696AA6-98DE-CB02-A519-584851008534}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B696AA6-98DE-CB02-A519-584851008534}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B696AA6-98DE-CB02-A519-584851008534}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|x64.Build.0 = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Debug|x86.Build.0 = Debug|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|x64.ActiveCfg = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|x64.Build.0 = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|x86.ActiveCfg = Release|Any CPU
+		{CDEF222D-7FE3-494E-AA82-A6E505583A4B}.Release|x86.Build.0 = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|x64.Build.0 = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Debug|x86.Build.0 = Debug|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|x64.ActiveCfg = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|x64.Build.0 = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|x86.ActiveCfg = Release|Any CPU
+		{F2455A4D-2DDD-42E0-B640-0D8065AC5637}.Release|x86.Build.0 = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|x64.Build.0 = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Debug|x86.Build.0 = Debug|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|x64.ActiveCfg = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|x64.Build.0 = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|x86.ActiveCfg = Release|Any CPU
+		{F1AC4F01-0AC1-44B0-9011-DF2468F5E309}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
